+++ src/npu/physiology.py
@@ -407,29 +407,3 @@ def update_upstream_db(src_cortical_area, src_neuron_id, dst_cortical_area, dst_
         runtime_data.upstream_neurons[dst_cortical_area][dst_neuron_id][src_cortical_area] = set()
     if src_neuron_id not in runtime_data.upstream_neurons[dst_cortical_area][dst_neuron_id][src_cortical_area]:
         runtime_data.upstream_neurons[dst_cortical_area][dst_neuron_id][src_cortical_area].add(src_neuron_id)
-
-
-def generate_plasticity_dict():
-    """
-    Extracts data from the genome and creates a dictionary of source cortical areas 
-    and their mapping destinations targeted for implementing neuroplasticity. 
-        
-        ex:    {
-                    'motor_memory': {
-                        'vision_memory',
-                        'auditory_memory'
-                    }
-                }
-
-    """
-    cortical_areas = runtime_data.genome['blueprint']
-    for area in cortical_areas:
-        for mapping_dst in cortical_areas[area]['cortical_mapping_dst']:
-            try:
-                if cortical_areas[area]['cortical_mapping_dst'][mapping_dst]['plasticity']:
-                    try:
-                        runtime_data.plasticity_dict[area].add(mapping_dst)
-                    except KeyError:
-                        runtime_data.plasticity_dict[area] = set(mapping_dst)
-            except KeyError:
-                pass
